Symbol "@feat.00" in windows PE obj must have storage class IMAGE_SYM_CLASS_STATIC
With IMAGE_SYM_CLASS_EXTERNAL the linker issues warnings about ignoring duplicate symbol definitions from other PE objs. The warnings are harmless but might be an issue in builds where linker warnings are treated as errors. Moreover, forcing the linker to ignore the warning (LNK4006) is not an option, because other symbol re-definitions might have undesired effects and thus should not go unnoticed.